requirements.txt
langchain
openai
python-dotenv
tiktoken
chromadb
matplotlib
squlite3
pyboxen